+++ admin-backend/admin-core/src/main/java/fun/yozora/admin/core/config/StpInterfaceImpl.java
@@ -1,9 +1,8 @@
 package fun.yozora.admin.core.config;
 
 import cn.dev33.satoken.stp.StpInterface;
-import cn.dev33.satoken.stp.StpUtil;
-import fun.yozora.admin.core.entity.Permission;
-import fun.yozora.admin.core.entity.Role;
+import fun.yozora.admin.domain.entity.Permission;
+import fun.yozora.admin.domain.entity.Role;
 import fun.yozora.admin.core.service.PermissionService;
 import fun.yozora.admin.core.service.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
