--- admin-backend/admin-core/src/main/java/fun/yozora/admin/core/entity/Role.java
@@ -1,38 +1,0 @@
-package fun.yozora.admin.core.entity;
-
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
-import lombok.Data;
-
-/**
- * 
- * @TableName role
- */
-@TableName(value ="role")
-@Data
-public class Role {
-    /**
-     * 
-     */
-    @TableId
-    private String roleId;
-
-    /**
-     * 角色编码
-     */
-    private String roleCode;
-
-    /**
-     * 
-     */
-    private String roleName;
-
-    /**
-     * 
-     */
-    private String description;
-
-
-}
