--- admin-backend/admin-core/src/main/java/fun/yozora/admin/core/entity/UserRole.java
@@ -1,6 +1,0 @@
-package fun.yozora.admin.core.entity;
-
-public class UserRole {
-    private String uid;
-    private String roleId;
-}
