--- admin-backend/admin-core/src/main/java/fun/yozora/admin/core/enums/MenuEnums.java
@@ -1,12 +1,0 @@
-package fun.yozora.admin.core.enums;
-
-public interface MenuEnums {
-    String BADGE_TYPE_DOT = "dot";
-    String BADGE_TYPE_NORMAL = "normal";
-
-    String BADGE_VARIANT_DEFAULT = "default";
-    String BADGE_VARIANT_DESTRUCTIVE = "destructive";
-    String BADGE_VARIANT_PRIMARY = "primary";
-    String BADGE_VARIANT_SUCCESS = "success";
-    String BADGE_VARIANT_WARNING = "warning";
-}
