--- admin-backend/admin-core/src/main/java/fun/yozora/admin/core/enums/UserStatus.java
@@ -1,7 +1,0 @@
-package fun.yozora.admin.core.enums;
-
-public enum UserStatus {
-    ACTIVE,
-    LOCKED,
-    DELETED;
-}
