--- admin-backend/admin-core/src/main/java/fun/yozora/admin/core/mapper/RoleMapper.java
@@ -1,19 +1,0 @@
-package fun.yozora.admin.core.mapper;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import fun.yozora.admin.core.entity.Role;
-import org.mapstruct.Mapper;
-
-/**
-* @author YOZORA
-* @description 针对表【role】的数据库操作Mapper
-* @createDate 2025-05-12 00:32:38
-* @Entity fun.yozora.admin.core.entity.Role
-*/
-public interface RoleMapper extends BaseMapper<Role> {
-
-}
-
-
-
-
