+++ admin-backend/admin-domain/src/main/java/fun/yozora/admin/domain/entity/Role.java
@@ -1,0 +1,38 @@
+package fun.yozora.admin.domain.entity;
+
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableField;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
+import lombok.Data;
+
+/**
+ * 
+ * @TableName role
+ */
+@TableName(value ="role")
+@Data
+public class Role {
+    /**
+     * 
+     */
+    @TableId
+    private String roleId;
+
+    /**
+     * 角色编码
+     */
+    private String roleCode;
+
+    /**
+     * 
+     */
+    private String roleName;
+
+    /**
+     * 
+     */
+    private String description;
+
+
+}
