+++ admin-backend/admin-domain/src/main/java/fun/yozora/admin/domain/entity/UserRole.java
@@ -1,0 +1,6 @@
+package fun.yozora.admin.domain.entity;
+
+public class UserRole {
+    private String uid;
+    private String roleId;
+}
