+++ admin-backend/admin-domain/src/main/java/fun/yozora/admin/domain/enums/BadgeType.java
@@ -1,0 +1,26 @@
+package fun.yozora.admin.domain.enums;
+
+public enum BadgeType {
+    DOT("dot"),
+    NORMAL("normal");
+
+    private final String value;
+
+    BadgeType(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public static BadgeType fromValue(String value) {
+        for (BadgeType type : values()) {
+            if (type.getValue().equals(value)) {
+                return type;
+            }
+        }
+        return NORMAL; // 默认值
+    }
+
+}
