+++ admin-backend/admin-domain/src/main/java/fun/yozora/admin/domain/enums/MenuEnums.java
@@ -1,0 +1,12 @@
+package fun.yozora.admin.domain.enums;
+
+public interface MenuEnums {
+    String BADGE_TYPE_DOT = "dot";
+    String BADGE_TYPE_NORMAL = "normal";
+
+    String BADGE_VARIANT_DEFAULT = "default";
+    String BADGE_VARIANT_DESTRUCTIVE = "destructive";
+    String BADGE_VARIANT_PRIMARY = "primary";
+    String BADGE_VARIANT_SUCCESS = "success";
+    String BADGE_VARIANT_WARNING = "warning";
+}
