+++ admin-backend/admin-domain/src/main/java/fun/yozora/admin/domain/enums/UserStatus.java
@@ -1,0 +1,7 @@
+package fun.yozora.admin.domain.enums;
+
+public enum UserStatus {
+    ACTIVE,
+    LOCKED,
+    DELETED;
+}
