+++ admin-backend/admin-repository/src/main/java/fun/yozora/admin/repository/mapper/RoleMapper.java
@@ -1,0 +1,18 @@
+package fun.yozora.admin.repository.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import fun.yozora.admin.domain.entity.Role;
+
+/**
+* @author YOZORA
+* @description 针对表【role】的数据库操作Mapper
+* @createDate 2025-05-12 00:32:38
+* @Entity fun.yozora.admin.domain.entity.Role
+*/
+public interface RoleMapper extends BaseMapper<Role> {
+
+}
+
+
+
+
