+++ admin-backend/admin-repository/src/main/resources/mapper/PermissionMapper.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="fun.yozora.admin.repository.mapper.PermissionMapper">
+
+    <select id="selectPermissionsByRoleId" resultType="fun.yozora.admin.domain.entity.Permission">
+        SELECT p.*
+        FROM permission p
+                 JOIN role_permission rp ON p.perm_id = rp.perm_id
+        WHERE rp.role_id = #{roleId}
+    </select>
+</mapper>
