+++ admin-backend/admin-repository/src/main/resources/mapper/UserMapper.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="fun.yozora.admin.repository.mapper.UserMapper">
+
+    <select id="selectByUsername" resultType="fun.yozora.admin.domain.entity.User">
+        SELECT *
+        FROM user
+        WHERE username = #{username}
+    </select>
+
+    <select id="selectByEmail" resultType="fun.yozora.admin.domain.entity.User">
+        SELECT *
+        FROM user
+        WHERE email = #{email}
+    </select>
+
+    <update id="updateLastLogin">
+        UPDATE user
+        SET last_login = #{lastLogin}
+        WHERE uid = #{uid}
+    </update>
+
+    <update id="updateIp">
+        UPDATE user
+        SET ip = #{ip}
+        WHERE uid = #{uid}
+    </update>
+</mapper>
