+++ admin-backend/admin-repository/src/main/resources/mapper/UserRoleMapper.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="fun.yozora.admin.repository.mapper.UserRoleMapper">
+
+    <select id="selectRolesByUserId" resultType="fun.yozora.admin.domain.entity.Role">
+        SELECT  r.role_id AS roleId,
+                r.description AS description,
+                r.role_code AS roleCode,
+                r.role_name AS roleName
+
+        FROM user_role ur
+        JOIN role r ON ur.role_id = r.role_id
+        WHERE ur.role_id = #{userId}
+    </select>
+
+</mapper>
