+++ admin-backend/admin-web/src/main/java/fun/yozora/admin/web/controller/AuthController.java
@@ -3,8 +3,8 @@ package fun.yozora.admin.web.controller;
 import cn.dev33.satoken.stp.StpUtil;
 import cn.dev33.satoken.util.SaResult;
 import fun.yozora.admin.web.dto.LoginDTO;
-import fun.yozora.admin.core.entity.User;
-import fun.yozora.admin.core.enums.UserStatus;
+import fun.yozora.admin.domain.entity.User;
+import fun.yozora.admin.domain.enums.UserStatus;
 import fun.yozora.admin.core.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
